Extract pattern availability hook
